dom/base/test/test_bug604592.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=604592
-->
<head>
  <title>Test for Bug 604592</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=604592">Mozilla Bug 604592</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">
/** Test for Bug 604592 **/
function testInStrictMode() {
  "use strict";
  try {
    document.body.prefix = "foo";
    ok(false, "Should not have reached this point");
  } catch (e) {
    ok(e instanceof TypeError, "Expected a TypeError");
  }
}

testInStrictMode();
document.body.prefix = "foo";
ok(document.body.prefix === null,
   "Expected strictly equal to null, got " + document.body.prefix);
</script>
</pre>
</body>
</html>
